This PR adds a `Metadata` struct that tracks the following information:

1. Start time
2. Stop time
3. Step time (time spent per step)
4. Step calls (how many times `step!` was called)
5. Allocations

It's created and passed as a keyword argument to `bundle_samples`, so it remains up to end-chain software to determine what to do (if anything) with the metadata.