@cpfiffer 2021-04-02
> I wonder if it would be better to track statistics in downstream packages (then one could also track other stuff or function calls etc.) or with callbacks (or a combination of both). I also wonder if it is necessary to add a special struct at all - couldn't we just keep track of these things in `mcmcsample` and then possibly forward them to `bundle_samples`? This would also get rid of the closures that I would like to avoid.

I think my idea here was to initialize something a little more extensible -- we could generally start adding more tracking fields in here to provide all kinds of profiling information to users that they cannot provide themselves because we sit on top of the sampling interface.

That said, I don't like the closure much either -- adding separate counter variables at the start is maybe preferable. It's much easier to reason about the performance rather than the closure weirdness.

> If the tracking leads to worse performance (some benchmarks would be interesting), it would be good if one could disable it. One should also be aware of that the statistics will also include compilation. I am also a bit curious if it would be more relevant in general to benchmark/track the whole `sample` call with `@timed` or some more advanced approach. Is it generally useful to exclude the allocations and times arising from the other parts of the `sample` call?

Same thing above here -- I want to provide something where we could glue more stuff to the inside to track some variables users might find interesting. I think we could add a separate field to track step times and total sampling times differently -- this could help people debug IO issues (`save!!`) or whatever.