@cpfiffer 2021-04-02
Just to add: the benefit of the struct is that someone can look at the docstring to see the available fields much more easily, so it integrates a little better with the dev experience. Not sure if that's worth the closure, though.