@devmotion 2021-04-02
Regardless of how the data is saved and forwarded to the user, I assumed it is not necessary to save it in a struct while running `sample` and closures are not needed. To be more precise, I thought the PR could be rewritten by modifying the `mcmcsample` implementations similar to
```julia
start = time()
step_time = 0.0
allocations = 0
(state, sample), _time, _allocated, _ = @timed step(...)
step_time += _time
allocations += _allocated

....

(state, sample), _step, _allocated, _ = @timed step(...)
step_time += _time
allocations += _allocated

...

stop = time()
duration = stop - start

stats = SamplingStatistics(start, stop, step_time, Ntotal, allocations)

bundle_samples(....; stats=stats)
```

In general, I think alternatively users could benchmark the whole sample run and compute the average time and allocations per sample from this information, or downstream packages could benchmark algorithm-specific things such as specific computationally expensive operations in the sampling algorithm. Additionally, downstream packages can even benchmark `step` by defining
```julia
function step(....)
    (sample, state), time, allocation, _ = @timed _step(...) # or whatever benchmark they are interested in
    return sample, withstats(state, (; time, allocation))
end

_step(...) = ... # the actual algorithm
````
So I still wonder what the (dis)advantages over the approach in this PR are.