@cpfiffer 2021-04-02
Ah, okay, I like that framework better.

As to the advantages, I think this is definitely one of those things we can introduce at the interface level and give to users for free if they want it -- it's simple enough to do and is something that I think we should consider offering broadly. 

The version you put above with no closures is ridiculously cheap from a computational perspective, so I don't think it has many disadvantages.