@cpfiffer 2021-04-06
Alrighty, I've pulled the closures out in favor of simple incrementers. 

I've also done a little benchmarking. I ran the following line on `master` and `csp/metadata`:

```julia
@benchmark chain = sample(MyModel(), MySampler(), 10_000; chain_type = MyChain, progress=false)
```

Here it is with all the `@timed` calls for this PR:

```
BenchmarkTools.Trial:
  memory estimate:  7.17 MiB
  allocs estimate:  188978
  --------------
  minimum time:     8.009 ms (0.00% GC)
  median time:      8.613 ms (0.00% GC)
  mean time:        9.145 ms (6.23% GC)
  maximum time:     17.446 ms (24.88% GC)
  --------------
  samples:          546
  evals/sample:     1
```

And on `master`:

```
BenchmarkTools.Trial:
  memory estimate:  1.98 MiB
  allocs estimate:  78996
  --------------
  minimum time:     2.066 ms (0.00% GC)
  median time:      2.189 ms (0.00% GC)
  mean time:        2.372 ms (4.61% GC)
  maximum time:     8.861 ms (36.89% GC)
  --------------
  samples:          2100
  evals/sample:     1
```

So, about a 4x increase in times. This isn't really that much of a complicated sampling process though, so I suspect most of this is just macro overhead that is very small relative to real computational costs. To test that, I ran it on some MH code.

This PR:

```
BenchmarkTools.Trial:
  memory estimate:  149.13 MiB
  allocs estimate:  3092902
  --------------
  minimum time:     209.231 ms (7.50% GC)
  median time:      215.551 ms (8.13% GC)
  mean time:        224.546 ms (11.45% GC)
  maximum time:     385.410 ms (45.19% GC)
  --------------
  samples:          23
  evals/sample:     1
```

`master`:

```
BenchmarkTools.Trial:
  memory estimate:  90.00 MiB
  allocs estimate:  1819232
  --------------
  minimum time:     140.753 ms (7.53% GC)
  median time:      141.976 ms (7.78% GC)
  mean time:        144.517 ms (8.01% GC)
  maximum time:     185.797 ms (7.83% GC)
  --------------
  samples:          35
  evals/sample:     1
```

It's an extra 60ms on average, which is absolutely god-awful. I'll just change this over to use start, stop, and duration only. Those `@timed` calls are brutal!