@devmotion 2021-04-06
I guess it might also help to make `SanplingStats` immutable (the values are not updated anymore) and to replace the type unions of the fields with concrete types.