@cpfiffer 2021-04-06
Updated this to use a more broad tracker that doesn't have a bajillion macro calls inside a loop. The current benchmarks are much more in line with `master`:

```
BenchmarkTools.Trial:
  memory estimate:  90.00 MiB
  allocs estimate:  1819235
  --------------
  minimum time:     125.679 ms (6.08% GC)
  median time:      126.867 ms (6.45% GC)
  mean time:        127.549 ms (6.66% GC)
  maximum time:     143.482 ms (5.85% GC)
  --------------
  samples:          40
  evals/sample:     1
```