@devmotion 2021-04-07
AFAICT the only problem is that on Julia 1.0 `@allocated` puts everything in a `let` block, i.e., you have to declare every variable that you want to return and use afterwards with `local`. I.e., similar to what you did with `state` you also have to add `local samples` etc.