@devmotion 2021-04-07
A completely different design that just came to my mind: Alternatively, instead of passing a struct to `bundle_samples`, one could log the statistics with a `@debug` message. These messages can contain keys with values and they are cheap since their computation and creating the message are skipped completely if debugging is not enabled. Downstream packages would have to handle these messages if they want to use the stats (e.g., with LoggingExtras) which might be a bit more inconvenient than passing it to `bundle_samples` - on the other hand, maybe it could be useful if not every implementation of `bundle_samples` has to implement how to deal with the stats in this way.