@cpfiffer 2021-04-08
I dunno, seems very low-cost to include -- I wouldn't want someone to catch rando log messages. Seems messy to me.

Maybe some context is important here. I wanted to add this because I've had some requests for things like ESS/minute, which requires time information, and I figured anyone who used AbstractMCMC might also want time information at some level.